Fix file descriptor race
If a pipe stream is created with read and write fd set to the same
value, the destroy method tries to close the same file descriptor
twice. However, the scheduler could choose another thread in between
that opens a file and gets the unused file descriptor re-assigned.
Hence, race condition.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@tele2.at>